refactor: Split into modules, refactor chunky methods with reduce()
