+++ lib/game_stats.rb
@@ -1,0 +1,51 @@
+module GameStats
+
+  def highest_total_score
+    max = @games.max_by { |game| game[:away_goals].to_i + game[:home_goals].to_i }
+    max[:away_goals].to_i + max[:home_goals].to_i
+  end
+
+  def lowest_total_score
+    min = @games.min_by { |game| game[:away_goals].to_i + game[:home_goals].to_i }
+    min[:away_goals].to_i + min[:home_goals].to_i
+  end
+
+  def percentage_home_wins
+    wins = @games.count { |game| game[:home_goals].to_i > game[:away_goals].to_i}
+    games = @games.count
+    (wins / games.to_f).round(2)
+  end
+
+  def percentage_visitor_wins
+    wins = @games.count { |game| game[:home_goals].to_i < game[:away_goals].to_i}
+    games = @games.count
+    (wins / games.to_f).round(2)
+  end
+
+  def percentage_ties
+    wins = @games.count { |game| game[:home_goals].to_i == game[:away_goals].to_i}
+    games = @games.count
+    (wins / games.to_f).round(2)
+  end
+
+  def count_of_games_by_season(team_id = false)
+    @games.reduce(Hash.new(0)) do |hash, game|
+      hash[game[:season]] += 1 if team_id == false || (team_id == game[:home_team_id || :away_team_id])
+      hash
+    end
+  end
+
+  def average_goals_per_game
+    (@games.sum { |game| game[:away_goals].to_f + game[:home_goals].to_f } / @games.count).round(2)
+  end
+
+  def average_goals_by_season
+    seasons = count_of_games_by_season
+    avg_arr = []
+    seasons.each do |season, count|
+      games_in_season = @games.find_all { |game| game[:season] == season }
+      avg_arr << ((games_in_season.sum { |game| game[:away_goals].to_i + game[:home_goals].to_i }) / count.to_f).round(2)
+    end
+    Hash[seasons.keys.zip(avg_arr)]
+  end
+end
